Setup for battleship
Install Jest for testing.
Install Babel
and its dependencies
according to the docs.

Configure Babel to target the
current version of Node in a
babel.config.js file.

Setup scripts to in package.json
to run Jest and another
to watch files and test on save.

Create gitignore and add
the node_modules and
coverage(for jest)
directories to it.