feat: Initialize Ship
Create a ship file and
its corresponding
test file.

Initialize a Ship class
that includes length,
hits and sunk instance
properties. The hits
property
would hold the number
of times the ship has
been
hit and the sunk
property will
hold infomation
of whether it has
been sunk of not.

Initialize a hit method
that increases the
hits property.

In the test file,
instantiate a new Ship
before very test
using the beforeEach()
hook ensures that
every test is
independent,
clean, and does
not depend on
previous tests'
leftovers.
Test the hit() method
to find out if
it increments the
hits property when
called. It
passes the test.