Implement updateSunkStatus
In the Ship class,
intialize a private
updateSunkStatus class
that updates the
sunk property of the
depending on the
length of the ship
and how many hits
it has taken.

Refactor the hits()
method. Remove the
return statement
and call the
updateSunkStatus()
so that the sunk
status is updated
after every hit.

In the test file,
check if the sunk
property is updated
as expected after
receiving a number
of hits.