Implement Gameboard and placeShip
Create a gameboard file and
its respective test file.

In gameboard file, initialize
a GameSquare class that
will represent one square on
the board. Set its ship instance
property to null and
its isHit property to false.
The ship property will hold
information if the the square
has a ship placed on it
and the isHit property will
hold information about if
the square has been hit.

Import the Ship class and
Initialize the GameBoard class.
Generate its 10 x 10 board
property (an array that
contains
10 arrays and each of
the 10 arrays have
10 elements)
by using an object
with a length property and
a nested Array.From method
and fill in all the squares
(elements of the 10 arrays)
with new instances of
the GameSquare.

Initialize a placeShip
method that takes in
the length of the ship
to be placed on the board
and the destination `x`
and `y` coordinates.
Create a new Ship with
the length and that
placeShip takes in.
Place the ship on
the coordinates
by setting the ship
property BoardSquare
that is at the coordinates
to the newly created Ship
and return true.

In the gameboard test file,
import GameBoard and
test placeShip. It passes
the test.