Setup for horizontal and vertical placement of ship
In GameBoard, add a direction
instance property because
the ships can be placed
either vertically or
horizontally.

Initialize a private
placeVertically method
that places the ship
vertically on the board.
The method takes in the
ship and its and
its destination
coordinates.
It loops through the
prospective vertical
position that the ship
would be placed and
checks if there's a
ship in any of the
positions (considering
the length of the
ship).
If no ship is in any
of the positions, it
loops through all those
positions, places
the ship
in those positions
according to the
ship's length and
returns true.

Implement similar
logic for the
private method
placeHorizontally
for horizontal
placement of ships.