Implement horizontal and vertical placement of ship
Initialize a withinBoard
method that takes in a
length and x and y
coordinates. The method
check if the length added
with each of the
each of the coordinates
are less than 10 because
it is a 10 x 10 board
and the indexes start
from 0. This would ensure
that the ship stays on
board (and considering
their length).

Refactor the placeShip
method to check if
the ship is withinBoard,
place the ship
according to the
the direction and
return what the
placeVertically or
placeHorizontally
method respectively.

Return false - the
direction is invalid.