Test vertical and horizontal placement of ships.
In gameboard, change the
name of the direction
instance variable to
placementDirection
and update all its
references.

Implement a
togglePlacementDirection
method that toggles
the placementDirection.

In the test file, test
vertical and horizontal
placement of ships
expecting
the ship object in
the start and end
coordinates where the
ships are placed.