Prepare to receive attack
In the ship file, have the
hit() method of the Ship
class to return a message.

In the GameBoard class
gameboard file, add a
missedHits instance
variable and set
it to an empty array.
It will hold the
coordinates of all
boardsquares that had
missed attacks. Add a
handleShipHit method
that takes in a
boardSquare and calls
the hit method on
the boardSquare's ship.

Also add a handleMissed
method that would take
in the coordinates of
a missed hit and push it
into the missedAttacks
array.