Implement and test receiveAttack
In the gameboard file,
initialize a receiveAttack
method in the GameBoard
class that
takes a pair of coordinates
and sends a hit to a ship
at the coordinates.

The method returns false
if the boardsquare has
been hit else it sets
the isHit property of the
boardsquare to be true.

If the said boardsquare
does not have a ship,
it pushes the coordinates
of the board to the
missedAttacks array else
it calls the handleShipHit
method on the boardsquare
and returns the result.

In the test file, test for
received attacks and
and recording missed
attacks.