Implement allShipsSunk
In the GameBoard class,
add a ships instance
variable and set it to
an empty array. This
will hold all the
ships after they are
placed.

Initialize a private
method addShip
that takes in
a ship and pushes it
into the ships array.

Call the addShip method
in the placeVertically
and placeHorizontally
on the shipToPlace
such that after the ship
is placed, it is pushed
into the ships array.

Initialize an allShipsSunk
method that calls the
array.every method to
return true is all ships
are sunk.