Create DOM controller module
Create a dom-controller file.
This will contoll DOM actions.

Initialize and export
a DomController
class. Initialize a
renderBoard method that
takes in a player and
a boardContainer and
renders the player's
board in the
boardContainer.

The method has a forEach
loop that loop through
the player's board and
for each row, it loops
through each element
(column), creates
gridCell element, adds
a class, adds
datasets x and y which
are the row index and
column index respectively
and appends it to the
boardContainer.

The x and y dataset can
be used to coordinates
later.