Begin HTML implementation
Create a HTML file and
populate it with boilerplate.
Create an index.js file
serve as an entry point
that initializes the
application. Create a
style.css file.

Set the src in the script
tag to index.js and set its
type as module so that it
will be handled as a
module. Link the style.css

In the body, set up the
board-container and
the human and computer
boards inside it.