Render boards
In index.js, import
the DomController and
the
Human and Computer
players.

Create new instances
of three of them.

Select the human and
computer board
containers and
render them by calling
the DomController's
renderBoard with the
right arguments.

Add styles in the
style.css so that the
boards' squares
are arranged in
10x10 grids.