Populate board with ships
In the gameboard file
in GameBoard class,
initialize a method
populateBoardWithDefaultShips
that populates the
board with default
ships. It calls
the placeShip with
various lengths
and coordinates and
calls
togglePlacementDirection
to toggle placement
direction as needed.

In the tests file, write
tests for the
populateBoardWithDefaultShips.

To test that the correct
number of ships were placed,
flatten the 2D board array
and filter all cells that
have ships (filter creates
a shallow copy with
the cells that passes
the filter condition).
Expect the
length of the total occupied
cells array to be the total
expected cells(based on the
ships' length that were placed
in
populateBoardWithDefaultShips).

Also test if ships were placed
in specific locations by testing
if they are vertically and
horizontally placed and
expecting a known ship's
length (based on a ship
placed in
populateBoardWithDefaultShips).