Store ship data
Rename the ships instance
variable and all its
references to
shipDataList because
it will be a list
of data objects, one
for each ship -
each ship's data list
will contain the
ship and its
coordinates

Rename the addShip and
all its references to
storeShipData because
it will handle the
storing of ships and
their coordinates to
the shipDataList.
Rename its parameter.

In placeVertically,
just before placing
the ship, create an
empty array - coordinates.
While placing the
ship, put the pairs
of coordinates into
an array and push the
array into the
coordinates array.

After the ship is
placed, put the
shipToPlace and
coodinates into
an object and call
storeShipData on
the object so
that they are pushed
into shipDataList.

Implement the same
for placeHorizontally

The ship's coordinates
will be used to
render the ships on
the DOM.