Prepare to render ships
Initialize a
getBoardContainer
that takes in
a player string and
returns the player's
board container.

Initialize a
getAllShipCoordinates
that takes in a
playerObject, loops
through its board's
shipListData and
returns an
array containing
all the coodinates
of all the ships on
the board.

Initialize a
getCellCoordinates
that takes in a cell
and returns its
coordinates from
its data attributes.