Apply class to cells
Initialize a
applyClassToMatchingCells
method
that takes in a cellsArray,
a coordinatesArray
and a className and adds
the className to the
cells in the cellsArray
that match the coordinates
in the coordinatesArray.

It loops through the
coordinates array and
for each coordinate,
it loops through the
cellsArray and for each
cell, it gets the cell's
coordinates by calling
getCellCoordinates.

It checks if the elements
of the current element
being looped through
in the coordinates array
corresponds with any
elements of any element
in the cells array
(each element in the
coordinates array
and the cells array are
a 2 element
array of coordinates)

If it finds a match,
it applies the className
to the class.

This will be used to fetch
coordinates from the player
object's board and update
the corresponding grid cells
on the player's board on
the DOM
with a visual representation
of the coordinates.