Render ships
In the
dom-controller file,
initialize a renderShips
method that takes in a
a player and a
playerobject and updates
the board with a visual
representation of the
ships.

The method gets the board
container by calling
the player on the
getBoardContainer. It
selects the cells
from the board
container.

Get the coordinates
of all the ships in
the board by calling
the getAllShipCoordinates
on the playerObject.

Apply the ship class
to all the cells whose
coordinates exist
in the getAllShipCoordinates
array by calling
applyClassToMatchingCells
on the board cells, all
the ships' coordinates
and the string ship.

Target and style
the ship class in the
style.css.