Generate random coordinates
To ensure that the
computer can make valid
moves,
(within the board)
initialize a
generateRandomCoordinates
method that returns an
array of 2 random numbers
from 0 to 9 (9 because it
is a 10 x 10 board and array
indexes start from 0).

Initialize a
generateRandomAttacksArray
that uses a for loop to
call generateRandomCoordinates
and
generate an array of 105
prospective random attack
coordinates
and return it.

This is because there are
there might be duplicate
coordinates generated and
if say the 100 (there are
100 squares on the board)
of the
coordinates generated
have been already attacked
(due to duplicate coordinates
generated)
the extra 5 can serve as
a slight buffer.