Launch next computer attack
Initialize a hasBeenAttacked
that takes in cellCoordinates
and a attackedCoordinatesArray.
It loops through the
attackedCoordinatesArray and
checks if the cellCoordinates
are in the array.

Initialize a
getNextComputerAttack that
returns the next computer
attack coordinates that
the computer has not
attacked before.

It dequeues an element
(currentElement)
from the random attacks
array, checks whether
it exists in the
missedAttacks array and
if it does, dequeue the
random attacks array
in a while loop and
continue the process
as long as it hasn't found
a coordinate
that has not been
attacked before and
random attacks
array has length.

Once any of the
conditions in the
while loop fails
(it has found a
coordinate that has not
been attacked or
the random attacks
array is empty
), it returns
the coordinates
(currentElement).

Initialize a
launchComputerAttack
that takes in a
playerObject and
calls the receiveAttack
method of the playerObject's
board.

It gets the attack
coordinates from
getNextComputerAttack
and returns the
result of calling
receiveAttack with
those coordinates.

This launches a
computer attack.