feat: Render missed attacks
First add new text
parameter to the
applyClassToMatchingCells
method and set its default
value to be an empty
string. After the class
is added to the matching
cells, set the
text content of the cell
to this new parameter.
This is so that it can
be possible to set the
text content of the
matching cells.

Initialize a
renderMissedAttacks
method. The method
takes in the player
and the playerObject,
gets the cells of
the player's board and
the playerObject's
gameboard's missed
attacks.
It then calls the
applyClassToMatchingCells
with these values, adding
and appropriate class
and a text content.

This renders the
playerObject's missedAttacks
on the player's board on
the DOM.