feat: Style missed attacks cells
First edit the style of
the grid template because
the text content of the
missed attacked cells
were causing the
missed attacked cells to
be uneven in height with
other cells. explictly
style the
grid template colums
and rows to giv them
equal width colums and
equal height rows.

Style the missed-cell
class' display,
font-size, weight and
background color.