Set board square coordinates
In the BoardSquare class
of the gameboard file
add a squareCoordinates
instance method. This
will hold the coordinates
of the square.

In the placeVertically and
placeHorizontally methods
of the GameBoard class,
after placing the ship,
set the coordinates
of the board square
accordingly.

In the test file, make
assertions to check
for the coordinates
of specified board squares
in the place vertically
and place horizontally
tests.