Get board cells
Initialize a private
method getBoardCells
that takes in a player
and a keyword. It gets
the board by calling the
getBoard method on
the player and the
runs a
querySelectorAll on
the player's board
with
the keyword and returns
the result.

Remove the boardContainer
and
allboardCells variables
from renderShips and
renderMissedAttacks and
instead, call
getBoardCells with the
player and the
respective keyword.