Render ships' hits
Initialize a
getHitShipsCoordinates
takes in a playerObject
that returns the
coordinates of all
ships that have been hit
on the player's board.

The method loops
through every row in
the player's board and
for each cell, if it
has a ship and it has
been hit, it pushes the
cell's coordinates
into an array and returns
the array.

Initialize a renderShipsHits
that takes in a player
and a player object. It
calls the getBoardCells
with the player and the
.ship keyword to get all
the cells that have the
class of ship (all
the cells that contains
ships).

Get the
hitShipsCoordinatesArray by
calling the
getHitShipsCoordinates
on the playerObject.

Call the
applyClassToMatchingCells
with the board cells that
contain ships, the
hitShipsCoordinatesArray,
an appropriate class
and text content.

This would apply the style
and give the text content
to the matching cells.