Add "X" to the hit ships with CSS
In styles.css, add an "X" mark
with CSS pseudo-elements
to the hit-ship class.
These pseudo-elements
are positioned absolutely
and rotated in opposite
directions (45 and -45 degrees)
to create the cross shape.

In the renderShipsHits of
the dom-comtroller, remove
the x text content when
the applying class
to matching cells because
the css has taken care of
that.