Style player ships differently
In renderShips, add a
className
parameter so that the
className
can be added dynamically
to all the matching ship
cells
with
applyClassToMatchingCells.
This is to be able to add a
class that is styled without
background colors to the
computer's ships as
background colors will
immediately reveal all
the computer's ships.

In index.js, call
renderShips() with
the respective
argument (className)
for
computerPlayer and
humanPlayer.

Update the style.css
to reflect these
newly added styles.

Now that the class names
for both players' ships
have changed and both end
in ship, update
renderHitShips to
getBoardCells of all the
cells that their class
ends with the "ship" word.