Setup game logic.
Create a game file
and in it, initialize
and export a Game class.

Setup 2 players and
the current player as
instance variables.
Initialize a
getCurrentPlayer
method that returns
the current player and
switchPlayerTurn that
switches the players.

Initialize a
playMove method that
takes in x and y
coordinates and
a handleBoardUpdate
callback function.
The method gets the
current player(which
is human) and saves the
result of launching
a human attack with
the x and y coordinates
on a variable. If the
move was a valid move,
its calls the callback
function to update the
computer's board
switches player turn,
proceeds to launch
the computer attack
and updates the
human board if
the computer move
is valid.