Handle board clicks
First update the
DomController class
to accept the players
as its paramaters so
that the Game class
can be called
Set instance variables
to these parameters

Import the Game class.
Set the game instance
variable to be a
new Instance of the
game class and pass in
these players parameters
into it.

Initialize a
handleBoardClicks as
an arrow function that
takes in an event.

handleBoardClicks
will be passed as in as
a callback to a click
event listener.
Initializing it as an
arrow function
ensures that the
scope of
`this` in it,
automatically
refers to the
DomController class
instance instead
of the element that
triggered the event.

handleBoardClicks
gets the row and
the column datasets
and returns if
the element clicked
on is not a grid
square.

If it is, call the
game's playMove method
with the row and column's
dataset and an
updateBoardDisplay
callback function.

Define the
updateBoardDisplay
as an arrow
function to
also preserve
its `this` scope
as the DomController
class instance
else it will lose
its `this` scope as
it is will be called
as a callback.