Update board display
Initialize
handlehumanBoardUpdate
and
handleComputerBoardUpdate.

Each of them gets the
corresponding player's
board, clear it,
render it again, render
its ships, its
missed attacks and
its ships hits.

In updateBoardDisplay,
call any of the
board updating methods
depending on the
player word that is
passed in.
(Remember playMoves
in the Game class
passes in an argument
to its callback function
depending
on which player's board
display is to be
updated.)