End game when ships sunk
Initialize a
checkAllShipsSunk
method
that takes in
the players objects,
and returns true if
any of the
players's objects
have had their ships
all sunk.
It checks the
allShipsSunk methods
of each player object's
gameboard.

In playGame, call
checkAllShipsSunk
to check before
the game starts
and after
the human
and computer player moves
if any of the player
have all their ships sunk
and if so, return and
end the game.