Separate human and computer moves
Have a different method
to handle each so that
the situation of a
player continuing
attack once he has hit
a ship until he misses
can be achieved.

Initialize a playHumanMove
property that first
checks if the all the
computer
player's ships are sunk
and returns a string if
they are.
Else it plays the move
(and saves the result
of the move to the
`result` variable),
updates the computer
board as necessary and
checks again if all
the computer's ships
have been sunk and
returns the string.
Else it returns
`result`.

Initialize
keepComputerAttacking
that first checks if
all the human ships have
been sunk, plays the
move(and saves the result
of the move to the
`result` variable),
updates the human
board as necessary
and checks again if
all human ships have
been sunk.

If the computer
hits a ship, it calls
keepComputerAttacking
repeatedly after every
2 seconds.
(Use setTimeout with an
arrow function to keep
the `this` context
of keepComputerAttacking
to the Game class)

Return `result`

Remove the
checkAllShipsSunk method
because it is no longer
needed.