Update handleBoardClicks
Update handleBoardClicks
to allow for continuous
attacking when a player
has hit a ship until
the player misses.

First rename the
selectedRow and
selectedColumn
variables
to xCoordinate and
yCoordinate
to better-fitting
names and update
their references
accordingly.

Launch the human attack
and proceed to
launch the computer
attack (keep computer
attacking) if the human
attack results in
`Missed`. Remember
that the
Game class'
keepComputerAttacking
method keeps calling
itself every 2 seconds
until its result
is "Missed"

Also set conditional
statements inbetween
the human and
and the
attacks that will
update
the DOM with the
winner, whenever
there is a winner.