Handle player turns
To prevent a player
from clicking his
board, implement a
system that handles
turns - making any
player's board it
is his turn to
play, inactive.

Create new instance
variables of both
players' board by
calling
the getBoardContainer
with the respective
strings.

Initialize a
disableHumanBoard that
adds a class which
disables the human
board. This will be
called at the start
of the game because
the human player
plays first.

Initialize a
toggleBoardState
that takes in
boardToEnable and
boardToDisable
and if boardToEnable
has the class that
disables board, it
removes it.
It adds it to the
boardToDisable

Update handleClicks
to toggleBoardState
with setTimeout.

Have it toggleBoard
state for computer to
play, 0.8 seconds
after the human plays.

Have it call the
computer attacking
section after
2 seconds (despite
keepComputerAttacking
advancing computer
attack every 2 seconds
if it does not miss
the previous attack)

Have it toggleBoard
back so that the
human can play
6 seconds after the
computer plays to
give some time for
the computer to
play at least 2
moves(if it hit the
previous one) before
toggling the board.
(All the setTimeouts
start counting
at the same time once
the method is called)

In the style.css,
add the class
that disables the
board. The class
sets the pointer
actions to none
and sets an opacity.

Call the DomController's
disableHumanBoard method
in the index.js so
the human board is
disabled when the game
starts as the human plays
first.