Change withinBoard logic
Change the logic for
withinBoard to
better serve the
intended purposes.

If the
placementDirection
is vertical,
it returns true if
x is less than 10,
y is less than 10
and the sum of x
and (length - 1)
is less than 10.

If the
placementDirection
is vertical,
it returns true if
x is less than 10,
y is less than 10
and the sum of y
and (length - 1)
is less than 10.

Call withinBoard
in placeShip, passing
in the required
arguments. Despite
that the logic of
withinBoard was not
right, the all ships
placed by
populateWithDefaultShips
were being placed
because withinBoard
was referenced in
placeShip, not
invoked as it should
have been.