Generate random placements
First rename the
generateRandomAttacksArray
to
generateRandomCoordinatesArray
and the randomAttacksArray
to randomCoordinatesArray
because
generateRandomCoordinatesArray
will be used to generate
random coordinates for
random placements.
Update its references

Initialize a
generateRandomPlacements
that takes in a length
argument.
It gets random
`coordinates` from
generateRandomCoordinatesArray,
randomizes the placement
direction,
and keeps taking each
coordinate from the
`coordinates` and
checking if the coordinate
is within the board.
It pushes all these
valid placements
and their placement
directions as an
object in
an array and returns
the array when there
are 10 valid placements.