Place ships randomly
Initialize
placeShipRandomly
that takes in a
length argument.
It gets the
`validPlacementsArray`
It sets `placementSuccess`
to false and
while it is still false,
check if the
validPlacementsArray
is empty and if it is,
generate new
validPlacementsArray.
Get each element
of the
validPlacementsArray,
unpack it to its
coordinates and
its direction.
Set the direction and
try to place a ship
in the coordinates
with the `length`
and the coordinates.

Return placementSuccess
(the result of placing
the ship).

This will keep looping
trying to place a ship
with `length`
`placementDirection`
on each coordinate
in the array and if the
`validPlacementsArray`
gets empty (due to
an inability to place
any of its previous
elements),
regenerate another
`validPlacementsArray`.

Rename
populateWithDefaultShips
to populateWithRandomShips
and call
placeShipRandomly
with different lengths
to populate the board
by randomly placing
the ships
with the specified
lengths.