Update getNextComputerAttack
Set previouslyAttacked
to true and
change the condition in
the while loop to just
be while
previouslyAttacked
is true.
Restructure it like
placeShipRandomly
by having generate
new randomAttacks
if randomAttacks
becomes empty at
any time.

This makes it so that
randomAttacks keeps
regenerating until
it gets a valid attack.

Reduce the number of
coordinates generated
in
generateRandomCoordinatesArray
to 10 because
the methods that use it,
keep generating new
coordinates until their
goals are met.

Change the condition
in the while loop of
to a more concise
alternative.