Reset board
Initialize a generateNewBoard
and have it return the 2D
array that `board` was
returning.
Set `board` to be the result
of generateNewBoard

Initialize a resetBoard
that generates a new
board, sets the placement
direction, empties the
missed attacks and
ship data list arrays
and populates the board
with random ships.

This method will be
used at the start
and end of a game
to reset the board.