Handle random ship placements
In the constructor, select
the announcement board,
the randomize ships
button, the start game
button and set them
as instance variables.
Disable the start game
button so that the
button will not be
active

Initialize a
handleRandomizeButtonClicks
that enables the startbutton,
resets each player's board
(re-populates it with ships
placed randonly)
and re-renders it.

Also update the announcement
board accordingly when the
computer wins.