Reset game
Add a reset button in
the index.html

Select the reset button
and the board container
in the constructor
and set them as instance
variables. Remove the
disable property from the
start button because
it has to be active at
the start of the game.

Initialize resetGame
and move the logic
from
handleRandomizeButtonClicks
that
resets the players'
board and update their
respective board displays
into it. Also have it
disable the reset button,
remove the event listener
from board container and
remove the disabled-board
class from any player's
board that has it.

In the
handleRandomizeButtonClicks,
call resetGame and
enable the start button.

Initialize a
handleResetButtonClicks
that takes in an event
calls resetGame,
disables the event.target
(basically disabling
itself) and enabling
the randomize button
and the start game
button.

handleResetButtonClicks
will be used as a callback
to the for the click
event listener
that will be attached
to the reset button.

The reset will be used
to reset the game
midway.