Handle start button clicks
Initialize a startGame,
that adds a click event
listener on the
reset button and passes
the
handleResetButtonClicks
as the callback function
so that users can
reset the game midway
by clicking on the
reset button

Initialize a
handleStartButtonClicks
that will be passed as
a callback for the
click event listener
on the start button.

handleStartButtonClicks
takes in an event,
disables the randomize
and start buttons
(because the game is
currently in session)
and disables the
reset button (so that
players can reset the
game midway)

It adds the disabled-board
class to the human board
(because it is the human
player's turn to play
first), adds the
a click event listener
on the board container
and passes the
handleBoardClicks
as a callback (so that
players can now make
their attacks on the
board) and calls
startGame.