Update announcement board
In the constructor,
set an
announcementText
instance variable
and set it as
the
announcement
board's text
content

Update the text content
of the announcement
board accordingly
depending on the
player that wins.

In initializeGame,
set the announcement
board's text content
accordingly. This
resets the announcement
board after a win and
after a reset.

Also update the
announcement board in
startGame.

In
handleHumanBoardUpdate
and
handleComputerBoardUpdate
used the
respective instance
variable used to
select the respective
boards instead of
selecting
the boards over again
in the methods.